Incrament

+++ LidarDemo/main.go
@@ -69,10 +69,10 @@ func RobotMainLoop(piProcessor *raspi.Adaptor, gopigo3 *g.Driver, lidarSensor *i
 				initialized = true
 			} else {
 				//Calculate where we want to go//
-				deltaY := previousDistance - lidarReading
-				deltaX := math.Sqrt( math.Pow( previousSpeed, 2 ) - math.Pow( deltaY ) )
+				deltaY := float64( previousDistance - lidarReading )
+				deltaX := math.Sqrt( math.Pow( float64( previousSpeed ), 2.0 ) - math.Pow( deltaY ) )
 				theta := math.Arctan( deltaY / deltaX )
-				speed := math.Sqrt( math.Pow( deltaX, 2 ) + math.Pow( deltaY, 2 ) )
+				speed := math.Sqrt( math.Pow( deltaX, 2.0 ) + math.Pow( deltaY, 2.0 ) )
 				//Calculate arc radius//
 				leftWheelLength := speed * math.Sin( theta )
 				rightWheelLength := speed * math.Cos( theta )
@@ -88,8 +88,8 @@ func RobotMainLoop(piProcessor *raspi.Adaptor, gopigo3 *g.Driver, lidarSensor *i
 				const RadiusConstant = math.Cos( Sidetheta ) * sideScaler * SideOffsetMagnitudeConstant */
 
 				//Calculate Dps for each wheel//
-				const LeftWheelDpsConstant = leftWheelLength / RobotWheelRadiusConstant
-				const RightWheelDpsConstant = rightWheelLength / RobotWheelRadiusConstant
+				const LeftWheelDpsConstant = int( leftWheelLength / RobotWheelRadiusConstant )
+				const RightWheelDpsConstant = int( rightWheelLength / RobotWheelRadiusConstant )
 				Move( gopigo3, LeftWheelDpsConstant, RightWheelDpsConstant )
 			}
 		} else if initialized == false {
